LLM Accuracy on Multiplication/requirements.txt
anthropic[bedrock]>=0.40.0
python-dotenv>=1.0.0
